TwitterClient -> TwitterApi + RestCall & RestListCall
These two Generics class allow callers of the TwitterApi to receive results
which are already typecast to the right type.  Also, the client can do the
job of creating the AsyncTask objects, and those tasks don't have to create
too many new objects.

Sadly, the RestCall still needs a .setClass() and .setItemClass() method to
construct the JavaType objects required by Jackson.  However, this all is
encapsulated by the TwitterApi client.